Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 int main(){
     int x,y,z;
     scanf("%d %d %d",&x,&y,&z);
-    float average = (x+y+z)/3;
+    float average = (float) (x+y+z)/3;
     scanf("%.2f",&average);
     printf("Average: %.2f",average);
     return 0;
